requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
matplotlib>=3.4.3
pyboy>=1.4.2
tensorboard>=2.6.0
tqdm>=4.62.3
pillow>=8.3.2